chen-seed: Completed data import functions
