+++ package.json
@@ -45,7 +45,7 @@
     "typescript": "^5.4.5"
   },
   "prisma": {
-    "seed": "ts-node prisma/seed.ts"
+    "seed": "ts-node prisma/seed"
   },
   "lint-staged": {
     "*.{ts,tsx}": [
