+++ src/utils/common.ts
@@ -1,0 +1,3 @@
+export const toLocaleDate = function(date: Date, type: 'en-CA' | 'en-US' | 'en-GB') : string{
+  return date.toLocaleDateString('en-ca');
+}
